Move link to header button
